requirements.txt
fritzconnection
requests
black
pylama
bandit
yamllint
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
python-dotenv
dateparser